% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\docType{data}
\name{add_data}
\alias{add_data}
\title{Simulated and Centered Extreme Air Temperature Data}
\format{
A numeric matrix with 30 rows and 15 columns. Rows represent years,
and columns represent different weather stations. Some values are missing (NA).

\describe{
\item{station1–station15}{Centered annual extreme air temperature values for each station.}
}
}
\source{
Simulated using the Generalized Extreme Value (GEV) distribution
fitted to observed data from Ribeirão Preto, Brazil.
}
\usage{
add_data
}
\description{
A simulated dataset of annual extreme air temperatures for 15 weather stations,
each centered by subtracting its sample mean. The values were generated from
a Generalized Extreme Value (GEV) distribution, with parameters fitted using
observed data from Ribeirão Preto, São Paulo, Brazil.
}
\details{
The dataset was created using the \code{dataset_add()} function.
}
\keyword{datasets}
