% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Heterogenety.R
\name{Add_Heterogenety}
\alias{Add_Heterogenety}
\title{Add_Heterogenety}
\usage{
Add_Heterogenety(dataset.add, rho, Ns)
}
\arguments{
\item{dataset.add}{A numeric matrix of air temperature data as that calculated by the dataset_add().}

\item{rho}{A single numeric value (constant) describing the average correlation among the sites.
It must be larger than -1.0 and lower than 1.0.}

\item{Ns}{Number of simulated groups of series.
Default is 100, but at least 500 is recommended.}
}
\value{
Hosking and Wallis' heterogeneity measure using an additive approach,
as proposed in Martins et al. (2022) \url{doi:10.1590/1678-4499.20220061}.
}
\description{
Add_Heterogenety
}
\examples{
rho <- 0.5
Ns <- 100
dataset.add <- add_data
Add_Heterogenety(dataset.add=dataset.add,rho = rho,Ns = Ns)
}
