% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_RegProb.R
\name{Add_RegProb}
\alias{Add_RegProb}
\title{Add_RegProb}
\usage{
Add_RegProb(quantiles, regional_pars, site_temp, n.year)
}
\arguments{
\item{quantiles}{A numeric vector with no missing data containing extreme air temperatures.}

\item{regional_pars}{A 6-column, 1-row \code{data.frame} or \code{matrix} as generated by \code{Reg_par()}.}

\item{site_temp}{A non-empty numeric vector or 1-column matrix of site air temperature data.}

\item{n.year}{A single integer giving the year index for which parameters are calculated.
Must be between 1 and \code{length(site_temp)}.}
}
\value{
A numeric matrix of cumulative probabilities (clamped to [0.0001, 0.9999])
associated with the input quantiles from the regional GEV distribution.
}
\description{
Add_RegProb
}
\examples{
quantiles <- c(37.8, 37.9, 38.0, 38.1, 38.2, 38.5, 39.0)
add.data <- Dataset_add(TmaxCPC_SP)
best.parms <- Best_model(add.data = add.data$add_data)
regional_pars <- Reg_par(best_model = best.parms$atsite.models)
Add_RegProb(
  quantiles     = quantiles,
  regional_pars = regional_pars,
  site_temp     = TmaxCPC_SP$Pixel_1,
  n.year        = 34
)
}
