% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_RegQuant.R
\name{Add_RegQuant}
\alias{Add_RegQuant}
\title{Regional Quantile Estimation for a Site Using Time-Varying GEV Parameters}
\usage{
Add_RegQuant(prob, regional_pars, site_temp, n.year)
}
\arguments{
\item{prob}{A numeric vector of probabilities strictly between 0 and 1, with no
missing values.}

\item{regional_pars}{A 1-row, 6-column \code{matrix} or \code{data.frame} containing regional
GEV parameters: location intercept, location slope, location slope (quadratic),
scale intercept, scale slope, and shape.}

\item{site_temp}{A numeric vector or 1-column matrix of site air temperature data.}

\item{n.year}{A single integer indicating the year index for which quantiles are
calculated. Must be between 1 and \code{length(site_temp)}.}
}
\value{
A named numeric vector of adjusted regional quantiles for the specified
year. Names are of the form \code{"Q90"}, `"Q95"``, etc.
}
\description{
Computes quantiles for a given site based on regional GEV parameters and
a time-varying location and scale model. Estimated quantiles are adjusted
by the site's mean temperature.
}
\examples{
prob <- c(0.90, 0.92, 0.93, 0.94, 0.95, 0.97, 0.99)
add.data <- Dataset_add(TmaxCPC_SP)
best.parms <- Best_model(add.data = add.data$add_data)
regional_pars <- Reg_par(best_model = best.parms$atsite.models)
Add_RegQuant(
  prob          = prob,
  regional_pars = regional_pars,
  site_temp     = TmaxCPC_SP$Pixel_1,
  n.year        = 34
)
}
