% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_sites.R
\docType{data}
\name{best_sites}
\alias{best_sites}
\title{At-site parameters of a time-varying generalized extreme value distribution}
\format{
\subsection{\code{best_sites}}{

A \code{dataframe} with 7 columns and 15 rows.

\describe{
\item{mu0}{1st factor of the location parameter for each site}
\item{mu1}{2nd factor of the location parameter}
\item{mu2}{3rd factor of the location parameter}
\item{sigma0}{1st factor of the scale parameter for each site}
\item{sigma1}{2nd factor of the scale parameter for each site}
\item{shape}{shape parameter for each site}
\item{size}{sample size for each site}
}
}
}
\source{
Generated by statistical simulation using \code{best_model()}.
}
\usage{
best_sites
}
\description{
At-site parameters of time-varying generalized extreme value distributions
from 15 simulated extreme air temperature series.
}
\keyword{datasets}
