% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_add.R
\name{dataset_add}
\alias{dataset_add}
\title{dataset_add}
\usage{
dataset_add(dataset)
}
\arguments{
\item{dataset}{A numeric matrix with extreme air temperature data from multiple sites.
The first column must contain the years, and the remaining columns contain
temperature data from each site.}
}
\value{
A \code{list} object with the following elements:
\describe{
\item{scaled_data}{A matrix of standardized (z-score) temperature values for each site.}
\item{reg_mean}{Mean temperature values for each site.}
\item{reg_sd}{Standard deviation of temperature values for each site.}
}
}
\description{
dataset_add
}
\examples{

dataset <- dataset
dataset_add <- dataset_add(dataset)

}
