% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_mult.R
\name{dataset_mult}
\alias{dataset_mult}
\title{dataset_mult}
\usage{
dataset_mult(dataset)
}
\arguments{
\item{dataset}{A numeric matrix with extreme air temperature data from multiple sites.
The first column must contain the years, and the remaining columns contain
temperature data from each site.}
}
\value{
A \code{list} object with the following elements:
\describe{
\item{divided_data}{A matrix of temperature values for each site divided
by their sample mean.}
\item{reg_mean}{Sample mean temperature values for each site.}
}
}
\description{
dataset_mult
}
\examples{

dataset <- dataset
dataset_mult <- dataset_mult(dataset)

}
