% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonlat_Tmax.R
\docType{data}
\name{lonlat_Tmax}
\alias{lonlat_Tmax}
\title{Geographical coordinates of the \code{TmaxCPC_SP} series}
\format{
\subsection{\code{lonlat_Tmax}}{

A data.frame with 2 columns and 10 rows.

\describe{
\item{lon}{Longitude in decinal degrees}
\item{lat}{Latitude in decinal degrees}
}
}
}
\source{
\url{https://psl.noaa.gov/}
}
\usage{
lonlat_Tmax
}
\description{
Geographical coordinates of the \code{TmaxCPC_SP} series
}
\keyword{datasets}
