% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mult_Heterogenety.R
\name{mult_Heterogenety}
\alias{mult_Heterogenety}
\title{Heterogenety measure calculated from the multiplicative approach}
\usage{
mult_Heterogenety(dataset.mult, rho, Ns)
}
\arguments{
\item{dataset.mult}{A numeric matrix of air temperature values divided by their sample means
for each site, as calculated by the dataset_mult() function.}

\item{rho}{A single numeric value (constant) describing the average correlation among the sites.
It must be larger than -1.0 and lower than 1.0.}

\item{Ns}{Number of simulated groups of series.
Default is 100, but at least 500 is recommended.}
}
\value{
Hosking and Wallis' heterogeneity measure using the multiplicative approach,
as proposed in Hosking and Wallis (1997) \url{doi:10.1017/CBO9780511529443}.
}
\description{
Heterogenety measure calculated from the multiplicative approach
}
\examples{
rho <- 0.5
Ns <- 100
divided_data <- divided_data
mult_Heterogenety(dataset.mult=divided_data,rho = rho,Ns = Ns)
}
