% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_pars.R
\docType{data}
\name{regional_pars}
\alias{regional_pars}
\title{Regional parameters of a time-varying generalized extreme value distribution}
\format{
\subsection{\code{regional_pars}}{

A \code{dataframe} with 5 columns and 1 row.

\describe{
\item{mu0}{1st factor of the location parameter for each site}
\item{mu1}{2nd factor of the location parameter}
\item{sigma0}{1st factor of the scale parameter for each site}
\item{sigma1}{2nd factor of the scale parameter for each site}
\item{shape}{shape parameter for each site}
}
}
}
\source{
Generated by statistical simulation using \code{reg_par()}.
}
\usage{
regional_pars
}
\description{
Regional parameters of time-varying generalized extreme value distributions
averaged from 15 simulated extreme air temperature series.
}
\keyword{datasets}
