% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TmaxCPC_SP.R
\docType{data}
\name{TmaxCPC_SP}
\alias{TmaxCPC_SP}
\title{Extreme air temperature data}
\format{
\subsection{\code{TmaxCPC_SP}}{

A data.frame with 14 columns and 34 rows.

\describe{
\item{Year}{Year}
\item{Pixel_1 to Pixel_13}{daily extremes for air temperature}
}
}
}
\source{
\url{https://psl.noaa.gov/}
}
\usage{
TmaxCPC_SP
}
\description{
Daily extremes for air temperature data sampled from
the block maxima approach. A block corresponds to one year.
The data belongs to The Climate Prediction Center - NOAA.
location of Ribeirao Preto, State of Sao Paulo in Brazil.
}
\keyword{datasets}
