% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Heterogeneity.R
\name{Add_Heterogeneity}
\alias{Add_Heterogeneity}
\title{Add_Heterogeneity}
\usage{
Add_Heterogeneity(dataset.add, rho, Ns)
}
\arguments{
\item{dataset.add}{A numeric matrix of air temperature data as that calculated by the dataset_add().}

\item{rho}{A single numeric value (constant) describing the average correlation among the sites.
It must be larger than -1.0 and lower than 1.0.}

\item{Ns}{Number of simulated groups of series.
Default is 100, but at least 500 is recommended.}
}
\value{
Hosking and Wallis' heterogeneity measure using an additive approach,
as proposed in Martins et al. (2022) \url{doi:10.1590/1678-4499.20220061}.
}
\description{
Add_Heterogeneity
}
\examples{
rho <- 0.51
Ns <- 500
add.data <- Dataset_add(TmaxCPC_SP)
Add_Heterogeneity(dataset.add=add.data$add_data,rho = rho,Ns = Ns)
}
