% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_RegProb.R
\name{Add_RegProb}
\alias{Add_RegProb}
\title{Add_RegProb}
\usage{
Add_RegProb(quantiles, reg_par, site_mean, site_sd, max_time)
}
\arguments{
\item{quantiles}{A numeric vector with no missing data with extreme air temperatures}

\item{reg_par}{A 6-column and 1-row \code{data.frame} as that generated by \code{reg_par()}.
\itemize{
\item 1st column is the mu0 parameter,
\item 2nd is the mu1 parameter,
\item 3rd is the mu2 parameter,
\item 4th is the sigma0 parameter,
\item 5th is the sigma1 parameter,
\item 6th is the shape parameter
}}

\item{site_mean}{A single number with site's average value for the air temperature data.}

\item{site_sd}{A single number with site's standard deviation for the air temperature data.}

\item{max_time}{A single number describing the number of the year that the time-varying parameters
should be calculated. For example, if the users need to calculated the parameters
for the first year max_time is set to 1 and the 30th year max_time is set to 30.}
}
\description{
Add_RegProb
}
\examples{
quantiles <- c(35,35.5,36,36.5,37,37.5,38,38.5,39,39.5,40)
reg_par <- regional_pars
RegProb <- Add_RegProb(quantiles=quantiles,
                  reg_par=reg_par,
                  site_mean=35.9,
                  site_sd=1.38,
                  max_time=30)
}
