% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_RegProb.R
\name{Add_RegProb}
\alias{Add_RegProb}
\title{Add_RegProb}
\usage{
Add_RegProb(quantiles, regional_pars, site_temp, n.year)
}
\arguments{
\item{quantiles}{A numeric vector with no missing data with extreme air temperatures.}

\item{regional_pars}{A 5-column and 1-row \code{data.frame} or \code{matrix} as that generated by \code{reg_par()}.}

\item{site_temp}{A vector or 1-column matrix with site's air temperature (original) data.}

\item{n.year}{A single number describing the number of the year for which parameters are calculated.}
}
\description{
Add_RegProb
}
\examples{
quantiles <- c(39.5,39.8,40.1,40.3,40.4,40.5,40.6,41.0,41.7)
regional_pars <- regional_pars
RegProb <- Add_RegProb(quantiles=quantiles,
                  regional_pars=regional_pars,
                  site_temp=dataset$station1,
                  n.year=30)
}
