% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_RegProb.R
\name{Add_RegProb}
\alias{Add_RegProb}
\title{Add_RegProb}
\usage{
Add_RegProb(quantiles, regional_pars, site_temp, n.year)
}
\arguments{
\item{quantiles}{A numeric vector with no missing data with extreme air temperatures.}

\item{regional_pars}{A 5-column and 1-row \code{data.frame} or \code{matrix} as that generated by \code{reg_par()}.}

\item{site_temp}{A vector or 1-column matrix with site's air temperature (original) data.}

\item{n.year}{A single number describing the number of the year for which parameters are calculated.}
}
\description{
Add_RegProb
}
\examples{
quantiles <- c(37.2, 37.5, 37.8, 37.9, 38.0, 38.1, 38.2, 38.5, 39.0)
add.data <- Dataset_add(TmaxCPC_SP)
best.parms <- Best_model(add.data=add.data$add_data)
regional_pars <- Reg_par(best_model = best.parms$atsite.models)
Add_RegProb(quantiles=quantiles,
                  regional_pars=regional_pars,
                  site_temp=TmaxCPC_SP$Pixel_1,
                  n.year=34)
}
