% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_RegQuant.R
\name{Add_RegQuant}
\alias{Add_RegQuant}
\title{Add_RegQuant}
\usage{
Add_RegQuant(prob, reg_par, site_mean, site_sd, max_time)
}
\arguments{
\item{prob}{A numeric vector with no missing data with the probabilities (between 0 and 1).}

\item{reg_par}{A 6-column and 1-row \code{data.frame} as that generated by \code{reg_par()}.
\itemize{
\item 1st column is the mu0 parameter,
\item 2nd is the mu1 parameter,
\item 3rd is the mu2 parameter,
\item 4th is the sigma0 parameter,
\item 5th is the sigma1 parameter,
}}

\item{site_mean}{A single number with site's average value for the air temperature data.}

\item{site_sd}{A single number with site's standard deviation for the air temperature data.}

\item{max_time}{A single number describing the number of years that the time-varying parameters
should be calculated.}
}
\description{
Add_RegQuant
}
\examples{
prob <- c(0.8,0.85,0.90,0.92,0.93,0.94,0.95,0.97,0.99)
best_model <- best_sites
reg_par <- reg_par(best_model=best_model)
RegQuant <- Add_RegQuant(prob=prob,
                  reg_par=reg_par,
                  site_mean=35.9,
                  site_sd=1.38,
                  max_time=30)
}
