% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Best_model.R
\name{Best_model}
\alias{Best_model}
\title{Time-varying parameters of the best fitted GEV model.}
\usage{
Best_model(add.data)
}
\arguments{
\item{add.data}{A numeric matrix of air temperature data as that calculated by the \code{dataset_add()} function.}
}
\value{
A list with:
\describe{
\item{best}{The model index (1 to 4) with the lowest total AIC across sites.}
\item{atsite.models}{A data frame of estimated time-varying parameters for the best model.}
}
}
\description{
This function fits four time-varying GEV models (with different assumptions on
non-stationarity in location and/or scale) to each temperature series, computes
the AIC of each model, and selects the best one per the lowest total AIC.
}
\details{
Model fitting is done using \code{ismev::gev.fit()}.
}
\examples{
add.data <- Dataset_add(TmaxCPC_SP)
best.parms <- Best_model(add.data=add.data$add_data)
}
