% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_model.R
\name{best_model}
\alias{best_model}
\title{Time-varying parameters of the best fitted GEV model.}
\usage{
best_model(add_data)
}
\arguments{
\item{add_data}{A numeric matrix of air temperature data subtracted by their sample means
for each site, as calculated by the dataset_add().}
}
\value{
A \code{list} object with the following elements:
\describe{
\item{best}{The best model among the six candidates.}
\item{atsite.models}{The time-varying parameters of the best model.}
}
}
\description{
Time-varying parameters of the best fitted GEV model.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
add_data <- add_data
best.parms <- best_model(add_data=add_data)
\dontshow{\}) # examplesIf}
}
