% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_model.R
\name{best_model}
\alias{best_model}
\title{Time-varying parameters of the best fitted GEV model.}
\usage{
best_model(temperatures)
}
\arguments{
\item{temperatures}{A numeric matrix of scaled (z-score) or divided air temperature data for each site,
as calculated by the Add_Discord() or Mult_Discord() functions.}
}
\value{
A \code{list} object with the following elements:
\describe{
\item{best}{The best model among the six candidates.}
\item{atsite.models}{The time-varying parameters of the best model.}
}
}
\description{
Time-varying parameters of the best fitted GEV model.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
add_data <- add_data
best.parms <- best_model(temperatures=add_data)
\dontshow{\}) # examplesIf}
}
