% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_add.R
\name{Dataset_add}
\alias{Dataset_add}
\title{dataset_add}
\usage{
Dataset_add(dataset)
}
\arguments{
\item{dataset}{A numeric matrix with extreme air temperature data from multiple sites.
The first column must contain the years, and the remaining columns contain
temperature data from each site.}
}
\value{
A \code{list} object with the following elements:
\describe{
\item{scaled_data}{A matrix of centered temperature values for each site.}
\item{reg_mean}{Mean temperature values for each site.}
}
}
\description{
dataset_add
}
\examples{

Dataset_add(TmaxCPC_SP)

}
