% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Time-varying parameters of a given GEV model.}
\usage{
fit_model(temperatures, model)
}
\arguments{
\item{temperatures}{A numeric matrix of scaled (z-score) or divided air temperature data for each site,
as calculated by the Add_Discord().}

\item{model}{A single interger number from 1 to 6 defining the GEV model.
May be provided by \code{best_model()}}
}
\value{
A \code{data.frame} with the time-varying parameters of the given model.
}
\description{
Time-varying parameters of a given GEV model.
}
\examples{
temperatures <- add_data
model <- 2
fit_model(temperatures=temperatures,model=model)
}
