% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Time-varying parameters of a given GEV model.}
\usage{
fit_model(add_data, model)
}
\arguments{
\item{add_data}{A numeric matrix of air temperature data subtracted by their sample means
for each site, as calculated by the dataset_add().}

\item{model}{A single interger number from 1 to 6 defining the GEV model.
May be provided by \code{best_model()}}
}
\value{
A \code{data.frame} with the time-varying parameters of the given model.
}
\description{
Time-varying parameters of a given GEV model.
}
\examples{
add_data <- add_data
model <- 2
fit_model(add_data=add_data,model=model)
}
