% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Time-varying parameters of a given GEV model.}
\usage{
fit_model(temperatures, model)
}
\arguments{
\item{temperatures}{A vector or single column matrix of air temperature data subtracted (or not) by its sample mean.}

\item{model}{A single integer number from 1 to 4 defining the GEV model.
May be provided by \code{best_model()}.}
}
\value{
A \code{data.frame} with the time-varying parameters of the given model.
}
\description{
Time-varying parameters of a given GEV model.
}
\examples{
temperatures <- dataset[,2:16]
model <- 2
fit_model(temperatures = temperatures, model = model)
}
