% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_par.R
\name{reg_par}
\alias{reg_par}
\title{reg_par}
\usage{
reg_par(best_model)
}
\arguments{
\item{best_model}{A 6-column \code{data.frame} as that generated by \code{best_model()}.
\itemize{
\item 1st column is the mu0 parameter,
\item 2nd is the mu1 parameter,
\item 3rd is the sigma0 parameter,
\item 4th is the sigma1 parameter,
\item 5th is the shape parameter,
\item 6th is the sample size.
}}
}
\value{
A \code{data.frame} with the regional parameters of a
time-varying generalized extreme value distribution
}
\description{
reg_par
}
\examples{
best_model <- best_sites
reg_par <- reg_par(best_model=best_model)
}
