% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg_parCI.R
\name{Reg_parCI}
\alias{Reg_parCI}
\title{Reg_parCI}
\usage{
Reg_parCI(temperatures, model, reg_par, n.boots)
}
\arguments{
\item{temperatures}{A numeric matrix of scaled (z-score) or divided air temperature data for each site,
as calculated by the Add_Discord() or Mult_Discord() functions.}

\item{model}{A single interger number from 1 to 6 defining the GEV model.
May be provided by \code{best_model()}}

\item{reg_par}{A 6-column and 1-row \code{data.frame} as that generated by \code{reg_par()}.
\itemize{
\item 1st column is the mu0 parameter,
\item 2nd is the mu1 parameter,
\item 3rd is the mu2 parameter,
\item 4th is the sigma0 parameter,
\item 5th is the sigma1 parameter,
\item 6th is the shape parameter
}}

\item{n.boots}{A single number describing the number of copies of the original dataset.
Whenever possible n.boots should be set to 999 (its default value),
as suggested by Burn (2003) <10.1623/hysj.48.1.25.43485> and
O’Brien and Burn (2014) <10.1016/j.jhydrol.2014.09.041>.}
}
\value{
A matrix containing the 95\% confidence intervals (lower and upper bounds)
of the time-varying parameter estimates. The spatial dependence between sites is preserved.
}
\description{
Reg_parCI
}
\examples{
temperatures <- add_data
model <- 2
reg_par <- regional_pars
n.boots <- 100
Reg_parCI(temperatures, model, reg_par, n.boots)
}
