% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site_parCI.R
\name{Site_parCI}
\alias{Site_parCI}
\title{Confidence intervals for site-specific GEV parameters}
\usage{
Site_parCI(atsite_temp, model, site_par, n.boots = 999)
}
\arguments{
\item{atsite_temp}{A numeric vector or single-column matrix of air temperature data.}

\item{model}{A single integer from 1 to 4 defining the GEV model.}

\item{site_par}{A 1-row, 5-column data.frame or matrix generated by Fit_model().}

\item{n.boots}{Number of bootstrap replicates. Values smaller than 100 are not allowed.
Defaults to \code{999}.}
}
\value{
A matrix containing the lower and upper 95\% confidence limits of
the estimated parameters.

A matrix containing the 95\% confidence intervals (lower and upper
bounds) of site-specific GEV parameters.
}
\description{
Confidence intervals for site-specific GEV parameters
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
temperatures <- TmaxCPC_SP$Pixel_1
model <- 2
site_par <- Fit_model(temperatures, model)

Site_parCI(
  atsite_temp = temperatures,
  model = model,
  site_par = site_par[1:5],
  n.boots = 100
)
\dontshow{\}) # examplesIf}
}
