% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site_parCI.R
\name{Site_parCI}
\alias{Site_parCI}
\title{Site_parCI}
\usage{
Site_parCI(atsite_temp, model, site_par, n.boots = 999)
}
\arguments{
\item{atsite_temp}{A vector or single-column matrix of air temperature data.}

\item{model}{A single integer number from 1 to 4 defining the GEV model.
May be provided by Best_model().}

\item{site_par}{A 5-column and 1-row data.frame or matrix as generated by Fit_model().
\itemize{
\item 1st column is the mu0 parameter,
\item 2nd is the mu1 parameter,
\item 3rd is the sigma0 parameter,
\item 4th is the sigma1 parameter,
\item 5th is the shape parameter
}}

\item{n.boots}{A single number describing the number of bootstrap replicates.
Whenever possible, n.boots should be set to 999 (default),
as suggested by Burn (2003) <10.1623/hysj.48.1.25.43485>
and O'Brien and Burn (2014) <10.1016/j.jhydrol.2014.09.041>.}
}
\value{
A matrix containing the 95\% confidence intervals (lower and upper bounds)
of the time-varying parameter estimates.
}
\description{
Site_parCI
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
temperatures <- TmaxCPC_SP$Pixel_1
model <- 2
site_par <- Fit_model(temperatures, model)

Site_parCI(
  atsite_temp = temperatures,
  model = model,
  site_par = site_par[1:5],
  n.boots = 100
)
\dontshow{\}) # examplesIf}
}
