% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site_parCI.R
\name{Site_parCI}
\alias{Site_parCI}
\title{Site_parCI}
\usage{
Site_parCI(atsite_temp, model, site_par, n.boots)
}
\arguments{
\item{atsite_temp}{A vector or single column matrix of air temperature data subtracted (or not) by its sample mean.}

\item{model}{A single interger number from 1 to 6 defining the GEV model.
May be provided by \code{best_model()}}

\item{site_par}{A 6-column and 1-row \code{data.frame}. may be obtained from that generated by \code{best.parms()}.
\itemize{
\item 1st column is the mu0 parameter,
\item 2nd is the mu1 parameter,
\item 3rd is the mu2 parameter,
\item 4th is the sigma0 parameter,
\item 5th is the sigma1 parameter,
\item 6th is the shape parameter
}}

\item{n.boots}{A single number describing the number of copies of the original data sample.
Whenever possible n.boots should be set to 999 (its default value),
as suggested by Burn (2003) <10.1623/hysj.48.1.25.43485> and
O’Brien and Burn (2014) <10.1016/j.jhydrol.2014.09.041>.}
}
\value{
A matrix containing the 95\% confidence intervals (lower and upper bounds)
of the time-varying parameter estimates.
}
\description{
Site_parCI
}
\examples{
atsite_temp <- add_data[,1]
model <- 2
site_par <- best_sites[1,1:6]
n.boots <- 100
Site_parCI(atsite_temp,model, site_par, n.boots)
}
